+++ src/assets/webpack/src/app/attributes/index.ts
@@ -4,3 +4,4 @@ export * from './json-editor';
 export * from './wysiwyg-editor';
 export * from './item';
 export * from './menu';
+export * from './menu-item-list';