+++ src/assets/webpack/src/app/attributes/list-line.ts
@@ -1,5 +1,8 @@
-import {bindable, customAttribute, ILogger, INode, resolve, IPlatform} from "aurelia";
+import {bindable, customAttribute, IDisposable, IEventAggregator, ILogger, INode, IPlatform, resolve} from "aurelia";
 import {ApiServices} from "../services/api-services";
+import {EEvents} from "../enums/events";
+import {IMoveMenuItem} from "../interfaces/menu-item";
+import {EMenuItemAction} from "../enums/menu";
 
 @customAttribute('cms-list-line')
 export class ListLine {
@@ -10,6 +13,7 @@ export class ListLine {
 
     public constructor(
         private readonly logger: ILogger = resolve(ILogger).scopeTo('ListLine'),
+        private readonly ea: IEventAggregator = resolve(IEventAggregator),
         private readonly element: HTMLElement = resolve(INode) as HTMLElement,
         private readonly apiService: ApiServices = resolve(ApiServices),
         private readonly platform:IPlatform = resolve(IPlatform)
@@ -20,11 +24,21 @@ export class ListLine {
     public attached()
     {
         this.logger.trace('attached');
-        this.addEvent();
     }
     public detached()
     {
-        this.logger.trace('attached');
+        this.logger.trace('detached');
+    }
+
+    public bound()
+    {
+        this.logger.trace('bound');
+        this.addEvent();
+    }
+
+    public unbinding()
+    {
+        this.logger.trace('unbinding');
         this.removeEvent();
     }
 
@@ -43,14 +57,13 @@ export class ListLine {
         if (this.buttonDelete) {
             this.buttonDelete.removeEventListener('click', this.onDelete);
         }
-
     }
 
     public onDelete = (event:Event) => {
         this.logger.trace('onDelete');
         event.preventDefault();
         if (this.buttonDelete) {
-            if (confirm('Attention !! vous allez supprimer définitivement cet utilisateur') === true) {
+            if (confirm('Attention !! vous allez supprimer définitivement cet élément ?') === true) {
                 const href:string = this.buttonDelete.href;
                 this.apiService.delete(href).then((response:any) => {
                     this.platform.taskQueue.queueTask(() => {