+++ src/assets/webpack/src/app/enums/api.ts
@@ -3,4 +3,5 @@ export enum EApi {
     IMPORT_ASYNC_UPLOAD = '/api/file/upload',
     IMPORT_ASYNC_PREVIEW = '/api/file/preview?name=__name__',
     IMPORT_ASYNC_DELETE = '/api/file/delete?name=__name__',
+    ITEM_MANAGE_MENU_ITEM = '/menu/{menuId}/manage-menu-items',
 }