+++ src/assets/webpack/src/app/interfaces/menu-item.ts
@@ -1,0 +1,11 @@
+/** Interface modal**/
+import {EMenuItemAction} from "../enums/menu";
+
+export interface IMenuItem {
+    sourceMenuItemId:number;
+    destMenuItemId:number;
+}
+export interface IMoveMenuItem {
+    name:string;
+    action:EMenuItemAction
+}